requirements.txt
numpy
opencv-python
Pillow
tensorflow
keras
fastapi
